// src/page/user/admin/enum/enum.ts
export enum OrderUserProperty {
	v_e = 'v_e',
	ci = 'ci',
	first_name = 'first_name',
	middle_name = 'middle_name',
	first_surname = 'first_surname',
	last_surname = 'last_surname',
	sex = 'sex',
	phone = 'phone',
	email = 'email',
}

export const textOrderUser = {
	v_e: 'Nacionalidad',
	ci: 'Cédula',
	first_name: 'Primer nombre',
	middle_name: 'Segundo nombre',
	first_surname: 'Primer apellido',
	last_surname: 'Segundo apellido',
	sex: 'Sexo',
	phone: 'Teléfono',
	email: 'Correo Electrónico',
};